modify data type correct and serial version is already fine

+++ c/loglikelihood.c
@@ -5,7 +5,7 @@ double loglikelihood(int **x, double mu[K][D], double *pi, double z[N][K]) {
     double result = 0;
     for (int n = 0; n < N; n++) {
         for (int k = 0; k < K; k++) {
-            int temp = 0;
+            double temp = 0;
 
             for (int i = 0; i < D; i++) {
                 temp = temp + x[n][i] * log(mu[k][i]) + (1 - x[n][i]) * log(1 - mu[k][i]);
